--- packages/mutations-ts/src/interface/dataSourceConfig.ts
@@ -1,4 +1,0 @@
-export interface DataSourceConfig{
-    graphqlEndpoint: string;
-    ipfs: any;
-}
